Now using getopt Also header files
