+++ src/Makefile.am
@@ -1,3 +1,3 @@
 bin_PROGRAMS = swap_wall
-swap_wall_SOURCES = swap_wall.c
+swap_wall_SOURCES = swap_wall.c read_file.c
 AM_CFLAGS = @GCONF_CFLAGS@ @GCONF_LIBS@ @GTK_CFLAGS@ @GTK_LIBS@
