+++ src/swap_wall.c
@@ -13,30 +13,31 @@
 #include <sys/stat.h>
 #include <string.h>
 #include <unistd.h>
-
+#include <getopt.h>
+#include "swap_wall.h"
 #include "read_file.h"
 
 char *current_dir;
 int time_gap = 10;
 char *style;
 
-gboolean set_wallpaper( const gchar *image_path, char *style);
-
-gboolean get_image(char *name){
-
+gboolean get_image(char *name)
+{
 	char *image_name = malloc(strlen(current_dir) + strlen(name) + 2);
 	strcpy(image_name, current_dir);
 	strcat(image_name, name);
 	return set_wallpaper(image_name, style);
 }
-void read_images(){
+
+void read_images()
+{
 	DIR *dir;
 	struct dirent *ent;
 	dir = opendir(current_dir);
-	
+
 	if(dir != NULL){
 		while((ent = readdir(dir)) != NULL){
-			
+
 			struct stat st_buf;
 			stat(ent -> d_name, &st_buf);
 			if(strcmp(ent->d_name, ".") != 0 && strcmp(ent->d_name, "..") != 0){
@@ -55,22 +56,26 @@ gboolean set_wallpaper( const gchar *image_path, char *style)
 
     client = gconf_client_get_default();
 
-    gconf_client_set_string( client, 
+    gconf_client_set_string( client,
 		"/desktop/gnome/background/picture_filename",
 		image_path,
 		NULL );
-       
-    gboolean result = gconf_client_set_string( client, 
-		"/desktop/gnome/background/picture_options", 
+
+    gboolean result = gconf_client_set_string( client,
+		"/desktop/gnome/background/picture_options",
         options,
         NULL);
     g_object_unref( G_OBJECT(client) );
     return result;
 }
-void kill_wall(int pid){
+
+void kill_wall(int pid)
+{
 	if(pid != 0) kill(pid, SIGKILL);
 }
-void start_wall(){
+
+void start_wall()
+{
 	current_dir = malloc(40);
 	style = malloc(20);
 	read_config(current_dir, style, &time_gap);
@@ -85,22 +90,55 @@ void start_wall(){
 		write_pid(pid);
 	}
 }
-int main(int argc, const char* argv[])
+
+int parse_arguments(int argc, char *argv[])
 {
-	printf("swap_wall:\n");
-	printf("------------\n\n");
-	if(argc > 1){
-		if(strcmp(argv[1], "--start") == 0){
-			start_wall();
-		}else if(strcmp(argv[1], "--restart") == 0){
-			kill_wall(read_pid());
-			start_wall();
-		}else if(strcmp(argv[1], "--stop") == 0){
-			kill_wall(read_pid());
-		}
-		printf("\n\n -------- done ! ----------\n\n");
-	}else{
-		printf("usage : \n\n swap_wall [--start|stop|restart] \n\n");
-	}
-	return 0;
+    int option_index = 0, c, i, j = 0;
+
+    while (1)
+    {
+        static struct option long_options[] = {
+            {"restart", no_argument, 0, 'r'},
+            {"stop", no_argument, 0, 's'},
+            {0, 0, 0, 0}};
+
+        c = getopt_long(
+                argc,
+                argv,
+                "rs",
+                long_options,
+                &option_index);
+
+        if (c == -1)
+            break;
+
+        switch (c)
+        {
+            case 'r':
+                kill_wall(read_pid());
+                start_wall();
+                break;
+
+            case 's':
+                kill_wall(read_pid());
+                break;
+
+            case '?':
+                exit(EXIT_FAILURE);
+                break;
+
+            default:
+                printf("Usage: swap_wall [--stop|restart]\n");
+                exit(EXIT_FAILURE);
+                break;
+        }
+    }
+    return 0;
+}
+
+int main(int argc, char* argv[])
+{
+    parse_arguments(argc, argv);
+    start_wall();
+    return 0;
 }
